requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
gunicorn==20.1.0
Flask-Login==0.6.2
Flask-Admin==1.6.0



# external utilities
requests>=2.0
qrcode[pil]>=7.0
Pillow>=9.0
captcha>=0.4
python-dotenv==1.0.1

python-bitcoinrpc>=1.0.0
